Make 'linear' dependency optional
Via 'nolinear' flag.
